+++ src/main/java/pretzel/dreamketcherbe/auth/jwt/JwtProvider.java
@@ -1,7 +1,6 @@
 package pretzel.dreamketcherbe.auth.jwt;
 
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtParser;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import lombok.extern.slf4j.Slf4j;
@@ -9,7 +8,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import io.jsonwebtoken.security.Keys;
-import pretzel.dreamketcherbe.member.entity.Role;
+import pretzel.dreamketcherbe.auth.repository.TokenProvider;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
